feat(ELE-2437): data/ split by nature (D2)
Separates four concerns that had been bundled in data/:

| Concern | New home |
|---|---|
| Engine abstractions | siege_utilities.engines |
| Statistics primitives | siege_utilities.data.statistics |
| Reference data | siege_utilities.reference |
| External data provider (RDH) | stays in data/ for now; moves to geo/providers/ under ELE-2438 |

## Moves

- data/dataframe_engine.py         -> engines/dataframe_engine.py
- data/cross_tabulation.py         -> data/statistics/cross_tabulation.py
- data/moe_propagation.py          -> data/statistics/moe_propagation.py
- data/naics_soc_crosswalk.py      -> reference/naics_soc_crosswalk.py
- data/sample_data.py              -> reference/sample_data.py

## Deprecation shims

Old paths still work but emit DeprecationWarning on import:
- siege_utilities.data.dataframe_engine
- siege_utilities.data.cross_tabulation
- siege_utilities.data.moe_propagation
- siege_utilities.data.sample_data
- siege_utilities.data.naics_soc_crosswalk

Remove in next minor release.

## Top-level data/__init__.py

Continues to re-export all public symbols for backward compatibility.
Callers using 'from siege_utilities.data import Engine, load_sample_data,
...' still work without deprecation warning.

## Updated internal callers

- siege_utilities/geo/boundary_sources.py (dataframe_engine)
- siege_utilities/survey/significance.py (cross_tabulation)
- tests/test_dataframe_engine.py + test_dataframe_engine_spatial.py
- tests/test_cross_tabulation.py
- tests/test_moe.py
- tests/test_sample_data.py
- tests/test_naics_soc.py

## Tests

tests/test_data_split_shims.py (10 tests) covers:
- New package layout imports cleanly
- Each deprecation shim fires DeprecationWarning
- data/__init__.py facade preserves public API

All 210 pre-existing data-related tests pass (3 unrelated fastparquet
environment failures). Survey tests green. Boundary sources import
verified.
